Copy main3.rb into main.rb
Delete main3.rb
Delete main2.rb
